rename `implied_bounds` module to `bounds`
